refactor(optimization): extract SVG optimization service and remove circular dependency
Breaks the circular dependency between extension.ts and svgEditorProvider.ts by
extracting shared optimization logic into a dedicated module.

Changes:
* move getSvgoPlugins and optimizeSvgDocument to svgOptimizationService.ts
* update extension.ts to import optimization helpers from the new service
* update svgEditorProvider.ts to consume optimizeSvgDocument from the service
* remove direct dependency from provider to extension module

This change is structural only and preserves existing behavior while improving
module boundaries and enabling independent evolution of core components.
